+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 4.13.0 (????? ???, ????; quick execution)
 
 * added `$ip()` function
+* added `$lastResult`, `$previousValue` and `$nextValue` variables
+* added `saveResultsToState` [setting](https://github.com/mkloubert/vs-script-commands#quick-execution-)
 
 ## 4.12.0 (May 2nd, 2017; REST API and cron jobs)